refactor(article): insert 페이지 생성, defineModel 적용
